+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Changelog
 
 v0.0.3-dev
 ----------
++ Add support for Ensembl transcripts on the links endpoint
 + Refactor and simplify the BedModel
 + Validate HGVS is valid before querying Variant Validator
 + Change the links endpoint to use POST and HGVS